split proposalResultItem into separate components
