+++ src/modules/components/proposal/proposalDataListItem/majorityVotingResult/index.ts
@@ -1,0 +1,1 @@
+export { MajorityVotingResult, type IMajorityVotingResultProps } from './majorityVotingResult';
