--- src/modules/components/proposal/proposalDataListItem/proposalDataListItemResult/index.ts
@@ -1,1 +1,0 @@
-export { ProposalDataListItemResult, type IProposalDataListItemResultProps } from './proposalDataListItemResult';
